+++ dist_node/pre_normalize.js
@@ -58,9 +58,9 @@ var splitUnpackList = (function(elements) {
     return ((indx < 0) ? [elements, null, []] : [elements.slice(0, indx), elements[indx], elements.slice((indx + 1))]);
 }),
     createUnpackList = (function(pre, mid, post) {
-        return concat(pre, ((mid && mid.id) ? SliceUnpack.create(null, mid.id, pre.length, post.length) : []), map(
-            (function(x, i) {
-                return RelativeUnpack.create(null, x, (post.length - i), (post.length + pre.length));
+        return concat(pre, ((mid && mid.id) ? SliceUnpack.create(null, mid.id, null, pre.length, post.length) : []),
+            map((function(x, i) {
+                return RelativeUnpack.create(null, x, null, (post.length - i), (post.length + pre.length));
             }), post));
     });
 peepholes.add("ArrayPattern", DOWN, (function(_) {