+++ dist_node/pseudo/pattern.js
@@ -1,26 +1,28 @@
 /*
- * THIS FILE IS AUTO GENERATED FROM 'lib/pseudo/pattern.kep'
+ * THIS FILE IS AUTO GENERATED from 'lib/pseudo/pattern.kep'
  * DO NOT EDIT
-*/
-"use strict";
+*/"use strict";
 var __o = require("khepri-ast")["node"],
     defineNode = __o["defineNode"],
     Node = __o["Node"],
     __o0 = require("khepri-ast")["pattern"],
     Pattern = __o0["Pattern"],
     RelativeUnpack, SliceUnpack;
-(RelativeUnpack = defineNode(Pattern, "RelativeUnpack", ["pattern"], ["index", "min"], (function(loc, pattern, index,
-    min) {
+(RelativeUnpack = defineNode(Pattern, "RelativeUnpack", ["pattern", "target"], ["index", "min"], (function(loc, pattern,
+    target, index, min) {
     var self = this;
     Node.call(self, loc);
     (self.pattern = pattern);
+    (self.target = target);
     (self.index = index);
     (self.min = min);
 })));
-(SliceUnpack = defineNode(Pattern, "SliceUnpack", ["pattern"], ["from", "to"], (function(loc, pattern, from, to) {
+(SliceUnpack = defineNode(Pattern, "SliceUnpack", ["pattern", "target"], ["from", "to"], (function(loc, pattern, target,
+    from, to) {
     var self = this;
     Node.call(self, loc);
     (self.pattern = pattern);
+    (self.target = target);
     (self.from = from);
     (self.to = to);
 })));